Fix a bug in filtering in the only-files option
There was a bug in the option --only-files, that it skips every file
in practice.

Filtering with only_files should happen only when downloading a file,
instead of downloading a folder. That's because in the context of
downloading a folder, it's not possible to get the final file name,
due to the recursive logic in download_folder().
As the most distinguishable provider names are included in the file
name, this mechanism should work fine.